requirements.txt
Django==4.2.9
requests>=2.25.1

django-filter>=23.2

django-simple-history>=3.4.0
django-import-export>=3.2.0